upload directory info - shared preferences, json
